220406 수업 - SampleGame - GameView size가 결정될 때 초기화되도록

+++ 220331_T_SampleGame/app/src/main/java/com/example/samplegame/Fighter.java
@@ -23,7 +23,7 @@ public class Fighter implements GameObject {
     private float dx, dy;  // 얼마나 이동할 것인지.
     private float tx, ty;  // 타겟의 좌표.
 
-    public Fighter(int x, int y) {
+    public Fighter(float x, float y) {
         this.x = x;  // 초기값 설정.
         this.y = y;
         this.tx = x;
